requirements.txt
# Web framework
flask==3.1.0
flask-cors==4.0.0

# Database
pymongo==4.6.1
bson==0.5.10
dnspython==2.4.2

# Environment and configuration
python-dotenv==1.0.0
requests>=2.26.0

# Include backend requirements
-r DevDocs/backend/requirements.txt

# Testing and development
pytest>=6.2.5
black>=21.9b0
isort>=5.9.3
flake8>=3.9.2

# Financial document processing
pdfplumber>=0.7.0
camelot-py>=0.10.1
tabula-py>=2.3.0
pandas>=1.3.0
matplotlib>=3.4.0
networkx>=2.6.0
unstructured>=0.5.0
opencv-python>=4.5.0
Pillow>=8.0.0
numpy>=1.20.0
scikit-learn>=1.0.0
PyPDF2>=2.0.0
pdf2image>=1.16.0
pytesseract>=0.3.8

# AI enhancement
requests>=2.26.0
langchain>=0.0.150
llama-index>=0.6.0
